py_server/requirements.txt
fastapi
uvicorn
onnx==1.18.0
onnxruntime==1.22.1
numpy
pillow 
python-multipart